Working load function.
