+++ practice_file.rb
@@ -1,0 +1,5 @@
+array_1 = [["stuff", "stuff2"], ["stuff3", "stuff4"]]
+
+array_1.each_with_object do |stuff|
+  puts stuff
+end
