+++ test/node_test.rb
@@ -1,7 +1,6 @@
 require 'minitest/autorun'
 require 'minitest/pride'
 require './lib/node'
-require 'pry'
 
 class NodeTest < Minitest::Test
   def test_node_exists
@@ -18,173 +17,4 @@ class NodeTest < Minitest::Test
     node = Node.new(12, "string")
     assert_equal "string", node.title
   end
-end
-
-class BinarySearchTreeTest < Minitest::Test
-  def test_Binary_search_tree_can_be_initialized
-    tree = BinarySearchTree.new
-    assert_instance_of(BinarySearchTree, tree)
-  end
-
-  def test_root_is_nil
-    tree = BinarySearchTree.new
-    assert_nil tree.root
-  end
-
-  def test_insert_method_sets_root
-    tree = BinarySearchTree.new
-    tree.insert(12, "String")
-    assert_instance_of(Node, tree.root)
-  end
-
-  def test_root_has_movie_score
-    tree = BinarySearchTree.new
-    tree.insert(12, "String")
-    assert_equal 12, tree.root.movie_score
-  end
-
-  def test_insert_method_assigns_left_node
-    tree = BinarySearchTree.new
-    tree.insert(12, "String")
-    tree.insert(10, "Second String")
-    assert_equal 10, tree.root.left.movie_score
-  end
-
-  def test_insert_method_assigns_right_node
-    tree = BinarySearchTree.new
-    tree.insert(12, "String")
-    tree.insert(10, "Second String")
-    tree.insert(15, "Third String")
-    assert_equal 15, tree.root.right.movie_score
-  end
-
-  def test_insert_method_assigns_left_left_node
-    tree = BinarySearchTree.new
-    tree.insert(12, "String")
-    tree.insert(10, "Second String")
-    tree.insert(15, "Third String")
-    tree.insert(9, "Fourth String")
-    assert_equal 9, tree.root.left.left.movie_score
-  end
-
-  def test_insert_method_assigns_left_left_left_node
-    tree = BinarySearchTree.new
-    tree.insert(12, "String")
-    tree.insert(10, "Second String")
-    tree.insert(15, "Third String")
-    tree.insert(9, "Fourth String")
-    tree.insert(8, "Fifth String")
-    assert_equal 8, tree.root.left.left.left.movie_score
-  end
-
-  def test_insert_method_assigns_right_right_node
-    tree = BinarySearchTree.new
-    tree.insert(12, "String")
-    tree.insert(15, "Third String")
-    tree.insert(18, "Thirs String")
-    assert_equal 18, tree.root.right.right.movie_score
-  end
-
-  def test_insert_method_assigns_right_right_node
-    tree = BinarySearchTree.new
-    tree.insert(12, "String")
-    tree.insert(15, "Third String")
-    tree.insert(10, "second String")
-    tree.insert(18, "Thirs String")
-    assert_equal 18, tree.root.right.right.movie_score
-  end
-
-  def test_insert_method_assigns_right_right_right_node
-    tree = BinarySearchTree.new
-    tree.insert(12, "String")
-    tree.insert(15, "Third String")
-    tree.insert(10, "second String")
-    tree.insert(18, "Thirs String")
-    tree.insert(20, "A different film")
-    assert_equal 20,  tree.root.right.right.right.movie_score
-  end
-
-  def test_insert_new_node_depth_returns_zero
-    tree = BinarySearchTree.new
-    assert_equal 0, tree.insert(12, "String")
-  end
-
-  def test_insert_new_node_increments_to_1
-    tree = BinarySearchTree.new
-    tree.insert(12, "String")
-    assert_equal 1, tree.insert(24, "Stringnig")
-  end
-
-  def test_insert_new_node_increments_to_2
-    tree = BinarySearchTree.new
-    tree.insert(12, "String")
-    tree.insert(24, "Stringnig")
-    assert_equal 2, tree.insert(34, "Stringy the third")
-  end
-
-  def test_insert_new_node_increments_correctly_left_and_right
-    tree = BinarySearchTree.new
-    tree.insert(12, "String")
-    tree.insert(24, "Stringnig")
-    tree.insert(10, "Stringy the third")
-    tree.insert(8, "Stringy the third")
-  end
-
-  def test_include_method
-    tree = BinarySearchTree.new
-    tree.insert(61, "Bill & Ted's Excellent Adventure")
-    tree.insert(16, "Johnny English")
-    tree.insert(92, "Sharknado 3")
-    tree.insert(50, "Hannibal Buress: Animal Furnace")
-    assert_equal true, tree.include?(16)
-    assert_equal false, tree.include?(72)
-    assert_equal false, tree.include?(100)
-  end
-
-  def test_depth_of_method
-    tree = BinarySearchTree.new
-    tree.insert(61, "Bill & Ted's Excellent Adventure")
-    tree.insert(16, "Johnny English")
-    tree.insert(92, "Sharknado 3")
-    tree.insert(50, "Hannibal Buress: Animal Furnace")
-    assert_equal 0, tree.depth_of(61)
-    assert_equal 1, tree.depth_of(16)
-    assert_equal 2, tree.depth_of(50)
-    # assert_equal nil, tree.depth_of(5)
-  end
-
-  def test_for_max
-    tree = BinarySearchTree.new
-    tree.insert(61, "Bill & Ted's Excellent Adventure")
-    tree.insert(16, "Johnny English")
-    tree.insert(92, "Sharknado 3")
-    tree.insert(50, "Hannibal Buress: Animal Furnace")
-    # assert_equal {"Sharknado 3"=>92}, tree.max
-    # this value is returned, but it's not right apparently?
-  end
-
-  def test_for_min
-    tree = BinarySearchTree.new
-    tree.insert(61, "Bill & Ted's Excellent Adventure")
-    tree.insert(16, "Johnny English")
-    tree.insert(92, "Sharknado 3")
-    tree.insert(50, "Hannibal Buress: Animal Furnace")
-    # assert_equal 16, tree.min
-  end
-
-  def test_sort_returns_two_values
-    skip
-    # This is really close to working. It's passing in a nil node.
-    tree = BinarySearchTree.new
-    tree.insert(61, "Bill & Ted's Excellent Adventure")
-    tree.insert(40, "Bill oiasdijo")
-    tree.insert(12, "iuhasdiuh")
-    assert_equal 12, tree.sort
-  end
-
-  def test_it_loads_a_text_file_and_increments_counter_per_file_loaded
-    tree = BinarySearchTree.new
-    assert_equal 98, tree.load('./lib/movies.txt')
-  end
-
 end
